movieQuest/requirements.txt
Flask==3.0.0
requests==2.32.3
urllib3<2
numpy<2
easyocr==1.7.0
opencv-python==4.8.1.78
Pillow==9.5.0